refactor into useJobsQuery
