+++ src/app/search/search.tsx
@@ -3,10 +3,10 @@
 import { Button, Divider, Paper, TextField } from "@mui/material"
 import { FormEvent } from "react"
 import { useForm } from "react-hook-form"
+import { useQueryParams } from "../../lib/hooks/useQueryParams"
 import { DutyDto } from "../api/duties/handler"
 import { LocationDto } from "../api/locations/handler"
 import { SkillDto } from "../api/skills/handler"
-import { useQueryParams } from "../useQueryParams"
 import { ClearanceFilter } from "./clearance-filter"
 import { DutyFilter } from "./duty-filter"
 import { LocationFilter } from "./location-filter"
